Perceptron Algorithm
Single neuron classifier using the Perceptron Learning Rule. This is a
two class problem with a two dimensional feature vector x. The data
sets provided below contain list of samples for each of the two
classes.  The fields in the data are: feature vector index, x[1],
x[2].Include a bias input x[0] and weight